requirements.txt
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
decorator==5.2.1
executing==2.2.0
idna==3.10
ipdb==0.13.13
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-bitcointx==1.1.5
requests==2.32.4
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.5.0
wcwidth==0.2.13
